Move tacotrom2 to prototype
