+++ test/torchaudio_unittest/models/tacotron2/model_test_cpu_test.py
@@ -1,19 +1,19 @@
 import torch
 
 from torchaudio_unittest.common_utils import PytorchTestCase
-from .model_test_impl import Tacotron2Encoder, Tacotron2Decoder, Tacotron2
+from .model_test_impl import Tacotron2EncoderTests, Tacotron2DecoderTests, Tacotron2Tests
 
 
-class TestTacotron2EncoderFloat32(Tacotron2Encoder, PytorchTestCase):
+class TestTacotron2EncoderFloat32(Tacotron2EncoderTests, PytorchTestCase):
     dtype = torch.float32
     device = torch.device('cpu')
 
 
-class TestTacotron2DecoderFloat32(Tacotron2Decoder, PytorchTestCase):
+class TestTacotron2DecoderFloat32(Tacotron2DecoderTests, PytorchTestCase):
     dtype = torch.float32
     device = torch.device('cpu')
 
 
-class TestTacotron2Float32(Tacotron2, PytorchTestCase):
+class TestTacotron2Float32(Tacotron2Tests, PytorchTestCase):
     dtype = torch.float32
     device = torch.device('cpu')
