+++ torchaudio/models/__init__.py
@@ -2,7 +2,6 @@ from .wav2letter import Wav2Letter
 from .wavernn import WaveRNN
 from .conv_tasnet import ConvTasNet
 from .deepspeech import DeepSpeech
-from .tacotron2 import _Tacotron2
 from .wav2vec2 import (
     Wav2Vec2Model,
     wav2vec2_base,
@@ -20,5 +19,4 @@ __all__ = [
     'wav2vec2_base',
     'wav2vec2_large',
     'wav2vec2_large_lv60k',
-    '_Tacotron2',
 ]
